unidad 1 ejercicio 3

+++ unidad1ejercicio2/ejercicio2.cpp
@@ -1,25 +1,25 @@
 #include <iostream>
 using namespace std;
 
-void ingresar(int &valor){
+void ingresar(float &valor){
     cout << "Ingresar valor: ";
     cin >> valor;
 }
 
-bool esValido(int a, int b){
+bool esValido(float a, float b){
     return (a!=b);
 }
 
-void mostrar(int valor1, int valor2){
+void mostrar(float valor1, float valor2){
     cout << "El mayor es: " << valor1 << " y el menor es: " << valor2 << endl;
 }
 
-void compararYmostrar(int a, int b){
+void compararYmostrar(float a, float b){
     (a>b) ? mostrar(a,b) : mostrar(b,a);
 }
 
 int main(){
-    int a, b;
+    float a, b;
     
     do{
         ingresar(a);
